+++ Sources/XcodeProj/Project/XcodeProj.swift
@@ -14,12 +14,16 @@ public final class XcodeProj: Equatable {
     /// Shared data.
     public var sharedData: XCSharedData?
 
+    /// User data.
+    public var userData: [XCUserData]
+
     // MARK: - Init
 
     public init(path: Path) throws {
         var pbxproj: PBXProj!
         var workspace: XCWorkspace!
         var sharedData: XCSharedData?
+        var userData: [XCUserData]
 
         if !path.exists { throw XCodeProjError.notFound(path: path) }
         guard let pbxprojPath = path.glob("*.pbxproj").first else {
@@ -35,9 +39,14 @@ public final class XcodeProj: Equatable {
         let sharedDataPath = path + "xcshareddata"
         sharedData = try? XCSharedData(path: sharedDataPath)
 
+        userData = XCUserData.path(path)
+            .glob("*.xcuserdatad")
+            .compactMap { try? XCUserData(path: $0) }
+
         self.pbxproj = pbxproj
         self.workspace = workspace
         self.sharedData = sharedData
+        self.userData = userData
     }
 
     public convenience init(pathString: String) throws {
@@ -49,10 +58,16 @@ public final class XcodeProj: Equatable {
     /// - Parameters:
     ///   - workspace: project internal workspace.
     ///   - pbxproj: project .pbxproj.
-    public init(workspace: XCWorkspace, pbxproj: PBXProj, sharedData: XCSharedData? = nil) {
+    ///   - sharedData: shared data
+    ///   - userData: user data
+    public init(workspace: XCWorkspace,
+                pbxproj: PBXProj,
+                sharedData: XCSharedData? = nil,
+                userData: [XCUserData] = []) {
         self.workspace = workspace
         self.pbxproj = pbxproj
         self.sharedData = sharedData
+        self.userData = userData
     }
 
     // MARK: - Equatable
@@ -60,7 +75,8 @@ public final class XcodeProj: Equatable {
     public static func == (lhs: XcodeProj, rhs: XcodeProj) -> Bool {
         lhs.workspace == rhs.workspace &&
             lhs.pbxproj == rhs.pbxproj &&
-            lhs.sharedData == rhs.sharedData
+            lhs.sharedData == rhs.sharedData &&
+            lhs.userData == rhs.userData
     }
 }
 
@@ -86,8 +102,8 @@ extension XcodeProj: Writable {
         try path.mkpath()
         try writeWorkspace(path: path, override: override)
         try writePBXProj(path: path, override: override, outputSettings: outputSettings)
-        try writeSchemes(path: path, override: override)
-        try writeBreakPoints(path: path, override: override)
+        try writeSharedData(path: path, override: override)
+        try writeUserData(path: path, override: override)
     }
 
     /// Returns workspace file path relative to the given path.
@@ -130,24 +146,48 @@ extension XcodeProj: Writable {
     /// - Parameter path: `.xcodeproj` file path
     /// - Returns: shared data path relative to the given path.
     public static func sharedDataPath(_ path: Path) -> Path {
-        path + "xcshareddata"
+        XCSharedData.path(path)
     }
 
-    /// Returns schemes folder path relative to the given path.
+    /// Writes shared data to the given path.
+    ///
+    /// - Parameter path: path to `.xcodeproj` file.
+    /// - Parameter override: if shared data should be overridden. Default is true.
+    /// - Parameter outputSettings: Controls the writing of various files.
+    ///   If false will throw error if shared data already exists at the given path.
+    public func writeSharedData(path: Path, override: Bool = true) throws {
+        try sharedData?.write(path: XCSharedData.path(path), override: override)
+    }
+
+    /// Writes user data to the given path.
+    ///
+    /// - Parameter path: path to `.xcodeproj` file.
+    /// - Parameter override: if user data should be overridden. Default is true.
+    /// - Parameter outputSettings: Controls the writing of various files.
+    ///   If false will throw error if user data already exists at the given path.
+    public func writeUserData(path: Path, override: Bool = true) throws {
+        for userData in userData {
+            try userData.write(path: XCUserData.path(path, userName: userData.userName), override: override)
+        }
+    }
+
+    /// Returns shared schemes folder path relative to the given path.
     ///
     /// - Parameter path: `.xcodeproj` file path
     /// - Returns: schemes folder path relative to the given path.
+    @available(*, deprecated, message: "use XCScheme.schemesPath(path:)")
     public static func schemesPath(_ path: Path) -> Path {
-        XcodeProj.sharedDataPath(path) + "xcschemes"
+        XCScheme.schemesPath(sharedDataPath(path))
     }
 
-    /// Returns scheme file path relative to the given path.
+    /// Returns shared scheme file path relative to the given path.
     ///
     /// - Parameter path: `.xcodeproj` file path
     /// - Parameter schemeName: scheme name
     /// - Returns: scheme file path relative to the given path.
+    @available(*, deprecated, message: "use XCScheme.path(path:schemeName)")
     public static func schemePath(_ path: Path, schemeName: String) -> Path {
-        XcodeProj.schemesPath(path) + "\(schemeName).xcscheme"
+        XCScheme.path(schemesPath(path), schemeName: schemeName)
     }
 
     /// Writes all project schemes to the given path.
@@ -157,34 +197,32 @@ extension XcodeProj: Writable {
     ///   If true will remove all existing schemes before writing.
     ///   If false will throw error if scheme already exists at the given path.
     public func writeSchemes(path: Path, override: Bool = true) throws {
-        guard let sharedData = sharedData else { return }
+        try sharedData?.writeSchemes(path: XCSharedData.path(path), override: override)
 
-        let schemesPath = XcodeProj.schemesPath(path)
-        if override, schemesPath.exists {
-            try schemesPath.delete()
-        }
-        try schemesPath.mkpath()
-        for scheme in sharedData.schemes {
-            try scheme.write(path: XcodeProj.schemePath(path, schemeName: scheme.name), override: override)
+        for userData in userData {
+            let userDataPath = XCUserData.path(path, userName: userData.userName)
+            try userData.writeSchemes(path: userDataPath, override: override)
         }
     }
 
-    /// Returns debugger folder path relative to the given path.
+    /// Returns shared debugger folder path relative to the given path.
     ///
     /// - Parameter path: `.xcodeproj` file path
     /// - Parameter schemeName: scheme name
     /// - Returns: debugger folder path relative to the given path.
+    @available(*, deprecated, message: "use XCDebugger.path(path:)")
     public static func debuggerPath(_ path: Path) -> Path {
-        XcodeProj.sharedDataPath(path) + "xcdebugger"
+        XCDebugger.path(XCSharedData.path(path))
     }
 
-    /// Returns breakpoints plist path relative to the given path.
+    /// Returns shared breakpoints plist path relative to the given path.
     ///
     /// - Parameter path: `.xcodeproj` file path
     /// - Parameter schemeName: scheme name
     /// - Returns: breakpoints plist path relative to the given path.
+    @available(*, deprecated, message: "use XCBreakpointList.path(path:)")
     public static func breakPointsPath(_ path: Path) -> Path {
-        XcodeProj.debuggerPath(path) + "Breakpoints_v2.xcbkptlist"
+        XCBreakpointList.path(debuggerPath(path))
     }
 
     /// Writes all project breakpoints to the given path.
@@ -194,13 +232,12 @@ extension XcodeProj: Writable {
     ///   If true will remove all existing debugger data before writing.
     ///   If false will throw error if breakpoints file exists at the given path.
     public func writeBreakPoints(path: Path, override: Bool = true) throws {
-        guard let sharedData = sharedData else { return }
+        let sharedDataPath = XcodeProj.sharedDataPath(path)
+        try sharedData?.writeBreakpoints(path: sharedDataPath, override: override)
 
-        let debuggerPath = XcodeProj.debuggerPath(path)
-        if override, debuggerPath.exists {
-            try debuggerPath.delete()
+        for userData in userData {
+            let userDataPath = XCUserData.path(path, userName: userData.userName)
+            try userData.writeBreakpoints(path: userDataPath, override: override)
         }
-        try debuggerPath.mkpath()
-        try sharedData.breakpoints?.write(path: XcodeProj.breakPointsPath(path), override: override)
     }
 }