+++ Sources/XcodeProj/Scheme/XCScheme.swift
@@ -130,3 +130,22 @@ public final class XCScheme: Writable, Equatable {
             lhs.wasCreatedForAppExtension == rhs.wasCreatedForAppExtension
     }
 }
+
+extension XCScheme {
+    /// Returns schemes folder path relative to the given path.
+    ///
+    /// - Parameter path: parent folder of schemes folder (xcshareddata or xcuserdata)
+    /// - Returns: schemes folder path relative to the given path.
+    static func schemesPath(_ path: Path) -> Path {
+        path + "xcschemes"
+    }
+
+    /// Returns scheme file path relative to the given path.
+    ///
+    /// - Parameter path: parent folder of schemes folder (xcshareddata or xcuserdata)
+    /// - Parameter schemeName: scheme name
+    /// - Returns: scheme file path relative to the given path.
+    static func path(_ path: Path, schemeName: String) -> Path {
+        XCScheme.schemesPath(path) + "\(schemeName).xcscheme"
+    }
+}